Massive changes to memory.
Memory is now virtually mapped so that it is very similar to how it is on the PS1. This way, we don't need to constantly check if the address being accessed is in a certain place, then map it ourselves, we let the OS take care of memory mapping.

Some problems - Probably won't compile for anything other than windows right now. Will fix later.

On windows, the memory must be aligned with the memory granularity. There are no guarentees about what that granularity is, so this may fail on runtime. I will have to implement a backup for this, which, again, I will do later.

Also still needs testing on x32 windows.