Cp0, Memory, and Timestamp stuff
Working on getting exceptions to work, getting Memory stuff back to the way it was before, and working on making timestamps more compatible with the dynarec. All of that is still kind of a WIP, but I think memory is good (maybe), and some exceptions are definetly hit. Timestamps are almost certainly not correct yet. Again, still a WIP.